openmanus_rl/agentbench/requirements.txt
numpy~=1.23.5
pydantic~=1.10.12
requests~=2.28.1
tqdm~=4.65.0
pyyaml~=6.0
jsonlines~=3.1.0
aiohttp~=3.8.4
uvicorn~=0.22.0
fastapi~=0.101.1
urllib3~=1.26.15
mysql-connector-python==8.0.33
docker==6.1.2
SPARQLWrapper
networkx~=2.8.4
anthropic~=0.4.1
fschat~=0.2.31
accelerate~=0.23.0
transformers~=4.34.0